Backend services

See merge request inf_se2_wise23_w.o.m.m/inf_se2_wise23_w.o.m.m!103

+++ src/main/java/de/hbrs/se2/womm/entities/Nutzer.java
@@ -23,7 +23,7 @@ public class Nutzer {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "nutzer_id")
-    private Integer nutzerId;
+    private long nutzerId;
 
     @Column(name = "nutzer_mail", nullable = false)
     private String nutzerMail;
